+++ src/PortalBase/Portal/Exception/DelegatingLoaderLoadException.php
@@ -8,13 +8,13 @@ final class DelegatingLoaderLoadException extends \Exception
 {
     private string $path;
 
-    public function __construct(string $path, ?\Throwable $previous = null)
+    public function __construct(string $path, int $code, ?\Throwable $previous = null)
     {
         $this->path = $path;
 
         parent::__construct(
             \sprintf('Exception when loading container service file from path %s', $path),
-            1674923696,
+            $code,
             $previous
         );
     }